refactor: Change Dir name (#153)
- req -> request
- res -> response
